Add cardano watcher into the processor
